requirements.txt
appdirs==1.4.0
enum34==1.1.6
pytz
scandir==1.3
setuptools
six==1.10.0
webdavclient>=1.0.8
